pyproject.toml
[tool.poetry]
name = "taskcontrol"
version = "0.1.0"
description = "Система контролей заданий на выпуск продукции"
authors = ["Никита Швейцер <shveitcer2023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mypy = "^1.8.0"
fastapi = "^0.109.2"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
flake8 = "^7.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
flake8-bugbear = "^24.2.6"
flake8-builtins = "^2.2.0"
pep8-naming = "^0.13.3"
flake8-variables-names = "^0.0.6"
flake8-import-order = "^0.18.2"
flake8-pyproject = "^1.2.3"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
pytest-asyncio = "^0.23.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "app"]
asyncio_mode = "auto"

[tool.flake8]
exclude = [
    ".git",
    ".env",
    ".env.example",
    "venv",
    "pycache",
    "mypy_cache",
    "pg_data",
    "alembic",
    "docker-compose.yaml",
    "alembic.ini",
    "LICENSE",
    "Dockerfile",
    "docs",
    "README.md"
]
max-line-length = 120
ignore = ["A003", "VNE003", "I100", "I201"]
max-complexity = 8